requirements.txt
# Core dependencies
fastapi==0.115.2
uvicorn==0.32.0
pydantic==2.7.4
python-dotenv==1.0.1

# HTTP clients
httpx==0.27.2

# External services
openai==1.52.0
pinecone>=5.4.0

# These are used by the memory manager
langchain-openai==0.2.3
langchain==0.3.4
langchain-pinecone==0.2.3

# For deployment
gunicorn==21.2.0

# Heroku
redis==5.2.1
groq==0.4.1
sendgrid>=6.11.0
twilio==9.3.5
markdown==3.6
nest_asyncio==1.6.0
sse_starlette==0.10.3
asyncio==3.4.3
python-multipart==0.0.20
asgiref==3.8.1
apscheduler==3.11.0
psycopg2-binary==2.9.10
tqdm==4.67.0
